Fix: Deferred rendering of CSS until page loaded
The links to the css files were enclosed in a noscript tag (that is only
used if the browser cannot run any scripts).

A new script was added that then inserts references to these styles right
after the page is loaded so that they can use the styles in the referred
css files, then removes the references.

Resolves: Google PageSpeed Insights recommendation to remove render-blocking
CSS