Fix: increased framerate for scrolling animation
This was the first attempt at increasing framerate by making
the updatePositions() function called by the scroll event listener
to be more efficient.  I did this by removing the variable assignment
from the loop and doing as much of the calculating as possible just
once before the loop.  I also counted the max number of pizzas on a
large high-res monitor and found it to be only 32, so substituted
this number for the entire list of pizza elements on the page.

Resolved: Slow and inefficient updating of pizza image
positions