Fix: Remove blank lines, move scripts to bottom
I moved scripts to the bottom of the page in an effort to
resolve the suggestion by Google PageSpeed Insights to
Prioritize visible content, but it doesn't seem to resolve
this.  It also doesn't effect the page negatively, so will
keep the scripts here for now.

I also removed some blank lines, although this saves space
will be back soon once I write comments.